fix(Network): manage include in ifdef to avoid compilation error on different plateform + handle connection closed by peer in server (when a client disconnect from the server)

+++ include/Network/Server.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 #include <map>
 #include <unordered_map>
@@ -41,10 +42,10 @@ class Server {
     uint16_t getPort() const { return _port; }
 
     // TCP
-    int acceptClient(Address& client_addr, uint currentTime);
+    int acceptClient(Address& client_addr, uint64_t currentTime);
 
     struct ClientInfo {
-        uint lastPacketTime;
+        uint64_t lastPacketTime;
         std::vector<uint8_t> input;
         std::vector<uint8_t> output;
     };
